feat: init test env
