+++ test.cpp
@@ -1,0 +1,3 @@
+#include "database.h"
+#include "boost/test/unit_test.hpp"
+
